requirements.txt
# Phase 1 — minimum runtime deps
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
gunicorn>=22,<23
whitenoise[brotli]>=6.7,<7
django-environ>=0.11,<0.12
django-cors-headers>=4,<5
requests>=2.32,<3
Pillow>=11,<12

# Storage (ArvanCloud S3) — activated via USE_S3_STORAGE env flag
django-storages[s3]>=1.14,<1.15
boto3>=1.35,<2

# Monitoring — activated via SENTRY_DSN env var
sentry-sdk>=2,<3

# --- Dev/test ---
pytest-django>=4,<5
factory-boy>=3,<4

# --- Added in later weeks (uncomment when needed) ---
# django-jalali>=7,<8
# django-unfold>=0.40,<1
# django-autoslug>=1.9,<2
# django-ratelimit>=4,<5
# django-phone-verify>=2.0,<3
# kavenegar>=1.1,<2

# --- Phase 2 upgrade packages ---
# psycopg[binary]>=3.2,<4
# django-redis>=5.4,<6
# celery>=5.4,<6
# django-celery-beat>=2.7,<3
